pull app.js' navigator toolbar into a Navigator component
