+++ views/types.js
@@ -2,11 +2,21 @@
 import {PropTypes} from 'react'
 import {Navigator} from 'react-native'
 
+export type RouteType = {
+  index: number,
+  id: string,
+  title: string,
+  backButtonTitle?: string,
+  onDismiss?: (any, any) => any,
+};
+
+export type NavStateType = {
+  routeStack: RouteType[],
+};
+
 export type TopLevelViewPropsType = {
-  navigator: typeof Navigator,
-  route: {
-    index: number,
-  },
+  navigator: Navigator,
+  route: RouteType,
 };
 
 export const TopLevelViewPropTypes = {
